refactor(server_stdio): extract tool-call lifecycle decorator to core/stdio_tool_lifecycle.py
Task 2 of docs/current/server-stdio-module-split.md. Pure extraction
(_log_tool_call decorator + protocol/compaction delivery state),
verbatim move verified byte-identical against git history. Removed
now-dead imports from server_stdio.py (json, functools, _debug,
claim_pending_compaction_prompt, ensure_marm_started,
maybe_auto_refresh, read_protocol_file, read_protocol_lite_file --
only consumer was this block).

Late-binding gotcha (flagged in the spec) confirmed and handled: the
decorator's body resolved several names from server_stdio.py's later
imports via Python's call-time name resolution. Now isolated in its
own module, it imports all of those itself. Verified with a real
end-to-end tool call (marm_log_show), not just an import-doesn't-crash
check -- confirmed logging, session init, and first-call protocol
injection all still fire correctly.

Updated tests/test_stdio_transport.py and tests/test_protocol_lite.py:
several tests monkeypatch or directly mutate protocol-delivery state
and its dependencies via module-qualified names (stdio.ensure_marm_started,
stdio._protocol_delivered, etc.) that lived directly on server_stdio
before this move. Repointed those patches to
core.stdio_tool_lifecycle, the code's new home -- no assertion or test
logic changed, only which module owns the patched name.

test_stdio_handles_mcp_initialize_and_exposes_tools (slow_stdio, real
subprocess spawn) flakes intentionally independent of this change --
verified by running it against the unmodified base commit, where it
also fails intermittently. Pre-existing environmental flakiness, not a
regression.
